% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.hivcd4.R
\name{expected_cd4}
\alias{expected_cd4}
\title{CD4 Progression Module}
\usage{
expected_cd4(
  method,
  cd4Count1,
  cd4Count2,
  male,
  age,
  ageInf,
  at,
  time.unit = 7
)
}
\description{
Function used in `hivvl_msm()` to handle CD4 count.
}
