% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.initialize.R
\name{InitErgmConstraint..attributes}
\alias{InitErgmConstraint..attributes}
\title{Initialize ERGM Constraint List for Faster Evaluation}
\usage{
InitErgmConstraint..attributes(lhs.nw, ...)
}
\arguments{
\item{lhs.nw}{Network object}

\item{...}{other arguments}
}
\description{
Comments out the computationally intensive constraint processing
             not needed by EMH models without constraints.
}
\keyword{initiation}
\keyword{msm}
\keyword{utility}
