% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.stitrans.xgc.R
\name{stitrans_msm_rand}
\alias{stitrans_msm_rand}
\title{STI Transmission Module}
\usage{
stitrans_msm_rand(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Stochastically simulates GC transmission given the current
             state of the edgelist. This version of the STI transmission
             function randomizes the transmission pathway responsible
             for a transmission event when multiple such events occur
             at a single anatomic site. Randomization addresses
             the depletion of susceptible anatomic sites that would
             occur if transmissions were assigned in the same order
             of transmission pathways every time, artificially depressing
             incidence due to transmission pathways that are processed
             later in the sequence.
}
\keyword{module}
\keyword{msm}
