% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_msm}
\alias{param_msm}
\title{Epidemic Model Parameters}
\usage{
param_msm(nwstats, hiv.test.int = c(301, 315), test.window.int = 21,
  tt.traj.B.prob = c(0.077, 0, 0.356, 0.567), tt.traj.W.prob = c(0.052,
  0, 0.331, 0.617), tx.init.prob = c(0.092, 0.127),
  tx.halt.prob = c(0.0102, 0.0071), tx.reinit.prob = c(0.00066,
  0.00291), max.time.off.tx.full.int = 520 * 7,
  max.time.on.tx.part.int = 52 * 15 * 7, max.time.off.tx.part.int = 520
  * 7, vl.acute.rise.int = 45, vl.acute.peak = 6.886,
  vl.acute.fall.int = 45, vl.set.point = 4.5, vl.aids.onset.int = 520
  * 7, vl.aids.int = 52 * 2 * 7, vl.fatal = 7, vl.full.supp = 1.5,
  vl.part.supp = 3.5, full.supp.down.slope = 0.25,
  full.supp.up.slope = 0.25, part.supp.down.slope = 0.25,
  part.supp.up.slope = 0.25, b.rate = 0.001/7, birth.age = 18,
  URAI.prob = 0.0082 * 1.09, UIAI.prob = 0.0031 * 1.09, acute.rr = 6,
  circ.rr = 0.4, cond.eff = 0.95, cond.fail = c(0.39, 0.21),
  circ.prob = c(0.874, 0.918), num.inst.ai.classes = 1,
  base.ai.main.BB.rate = 0.22, base.ai.main.BW.rate = 0.22,
  base.ai.main.WW.rate = 0.22, base.ai.pers.BB.rate = 0.14,
  base.ai.pers.BW.rate = 0.14, base.ai.pers.WW.rate = 0.14,
  ai.scale.BB = 1.31, ai.scale.BW = 1, ai.scale.WW = 0.77,
  cond.main.BB.prob = 0.21, cond.main.BW.prob = 0.21,
  cond.main.WW.prob = 0.21, cond.pers.always.prob = 0.216,
  cond.pers.BB.prob = 0.26, cond.pers.BW.prob = 0.26,
  cond.pers.WW.prob = 0.26, cond.inst.always.prob = 0.326,
  cond.inst.BB.prob = 0.27, cond.inst.BW.prob = 0.27,
  cond.inst.WW.prob = 0.27, cond.always.prob.corr = 0.5,
  cond.rr.BB = 0.71, cond.rr.BW = 1, cond.rr.WW = 1.6,
  prep.start = Inf, prep.coverage = 0,
  prep.adhr.dist = reallocate_pcp(reall = 0.784 - 0.784),
  prep.class.hr = c(0.69, 0.19, 0.02), prep.discont.rate = 1 -
  (2^(-1/780)), prep.tst.int = 90, prep.risk.int = 182,
  prep.risk.reassess.method = "year", rcomp.prob = 0.41,
  rcomp.adh.groups = 2:3, rgc.tprob = 0.428, ugc.tprob = 0.35,
  rct.tprob = 0.231, uct.tprob = 0.205, rgc.sympt.prob = 0.16,
  ugc.sympt.prob = 0.9, rct.sympt.prob = 0.14, uct.sympt.prob = 0.58,
  rgc.ntx.int = 205.8, ugc.ntx.int = 205.8, gc.tx.int = 2 * 7,
  rct.ntx.int = 265.1, uct.ntx.int = 265.1, ct.tx.int = 2 * 7,
  gc.sympt.prob.tx.B = 0.86, gc.sympt.prob.tx.W = 0.96,
  ct.sympt.prob.tx.B = 0.72, ct.sympt.prob.tx.W = 0.85,
  gc.asympt.prob.tx.B = 0.1, gc.asympt.prob.tx.W = 0.19,
  ct.asympt.prob.tx.B = 0.05, ct.asympt.prob.tx.W = 0.1,
  prep.sti.screen.int = 182, prep.sti.prob.tx = 1,
  prep.continue.stand.tx = TRUE, sti.cond.eff = 0.95,
  sti.cond.fail.B = 0.39, sti.cond.fail.W = 0.21, hiv.rgc.rr = 2.78,
  hiv.ugc.rr = 1.73, hiv.rct.rr = 2.78, hiv.uct.rr = 1.73,
  hiv.dual.rr = 0.2, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats_msm}}.}

\item{hiv.test.int}{Mean intertest interval in days for black/white MSM
(vector of length 2).}

\item{test.window.int}{Length of the HIV test window period in days.}

\item{tt.traj.B.prob}{Proportion of black MSM who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tt.traj.W.prob}{Proportion of white MSM who enter into the four
testing/treatment trajectories, as defined above.}

\item{tx.init.prob}{Probability per time step that a black/white MSM who has
tested positive will initiate treatment (vector of length 2).}

\item{tx.halt.prob}{Probability per time step that a black/white MSM who is
currently on treatment will halt treatment (vector of length 2).}

\item{tx.reinit.prob}{Probability per time step that a black/white MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment (vector of length 2).}

\item{max.time.off.tx.full.int}{Number of days off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of days on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of days off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in days.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{b.rate}{Rate at which MSM enter the population.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{cond.eff}{Relative risk of HIV infection from anal sex when a condom is
used properly (biological efficacy).}

\item{circ.prob}{Probablity that a black/white new arrival in the population
will be circumcised (vector of length 2).}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{base.ai.main.BB.rate}{Expected coital frequency in black-black main
partnerships (acts per day).}

\item{base.ai.main.BW.rate}{Expected coital frequency in black-white main
partnerships (acts per day).}

\item{base.ai.main.WW.rate}{Expected coital frequency in white-white main
partnerships (acts per day).}

\item{base.ai.pers.BB.rate}{Expected coital frequency in black-black casual
partnerships (acts per day).}

\item{base.ai.pers.BW.rate}{Expected coital frequency in black-white casual
partnerships (acts per day).}

\item{base.ai.pers.WW.rate}{Expected coital frequency in white-white casual
partnerships (acts per day).}

\item{ai.scale.BB}{Relative scaler for all BB act rates.}

\item{ai.scale.BW}{Relative scaler for all BW act rates.}

\item{ai.scale.WW}{Relative scaler for all WW act rates.}

\item{cond.main.BB.prob}{Probability of condom use in a black-black main
partnership.}

\item{cond.main.BW.prob}{Probability of condom use in a black-white main
partnership.}

\item{cond.main.WW.prob}{Probability of condom use in a white-white main
partnership.}

\item{cond.pers.always.prob}{Fraction of men in casual partnerships who always
use condoms in those partnerships.}

\item{cond.pers.BB.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-black casual partnerships.}

\item{cond.pers.BW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-white casual partnerships.}

\item{cond.pers.WW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a white-white casual partnerships.}

\item{cond.inst.always.prob}{Fraction of men in instant partnerships who always
use condoms in those partnerships.}

\item{cond.inst.BB.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-black one-off partnerships.}

\item{cond.inst.BW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-white one-off partnerships.}

\item{cond.inst.WW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a white-white one-off partnerships.}

\item{cond.always.prob.corr}{Correlation coefficient for probability of always
using condoms in both casual and one-off}

\item{cond.rr.BB}{Condom probability scaler for black-black partnerships for
model calibration purposes.}

\item{cond.rr.BW}{Condom probability scaler for black-white partnerships for
model calibration purposes.}

\item{cond.rr.WW}{Condom probability scaler for white-white partnerships for
model calibration purposes.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.class.hr}{The hazard ratio for infection per act associated with each
level of adherence (from Grant).}

\item{prep.discont.rate}{Rate of random discontinuation from PrEP.}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in days.}

\item{prep.risk.reassess.method}{Method for determining risk-based discontinuation
of PrEP, with \code{"none"} for no discontinuation, \code{"inst"} for
reassessment every time step, and \code{"year"} for reassessment at yearly
HIV diagnostic testing visits.}

\item{rcomp.prob}{Level of risk compensation from 0 to 1, where 0 is no risk
compensation, 0.5 is a 50% reduction in the probability of condom use
per act, and 1 is a complete cessation of condom use following PrEP
initiation.}

\item{rcomp.adh.groups}{PrEP adherence groups for whom risk compensation
occurs, as a vector with values 1, 2, 3 corresponding to
low adherence, medium adherence, and high adherence to PrEP.}

\item{rgc.tprob}{Probability of rectal gonorrhea infection per act.}

\item{ugc.tprob}{Probability of urethral gonorrhea infection per act.}

\item{rct.tprob}{Probability of rectal chlamydia infection per act.}

\item{uct.tprob}{Probability of urethral chlamydia infection per act.}

\item{rgc.sympt.prob}{Probability of symptoms given infection with rectal
gonorrhea.}

\item{ugc.sympt.prob}{Probability of symptoms given infection with urethral
gonorrhea.}

\item{rct.sympt.prob}{Probability of symptoms given infection with rectal
chlamydia.}

\item{uct.sympt.prob}{Probability of symptoms given infection with urethral
chlamydia.}

\item{rgc.ntx.int}{Average duration in days of untreated rectal gonorrhea.}

\item{ugc.ntx.int}{Average duration in days of untreated urethral gonorrhea.}

\item{gc.tx.int}{Average duration in days of treated gonorrhea (both sites).}

\item{rct.ntx.int}{Average in days duration of untreated rectal chlamydia.}

\item{uct.ntx.int}{Average in days duration of untreated urethral chlamydia.}

\item{ct.tx.int}{Average in days duration of treated chlamydia (both sites).}

\item{gc.sympt.prob.tx.B}{Probability of treatment for symptomatic gonorrhea
for Black men.}

\item{gc.sympt.prob.tx.W}{Probability of treatment for symptomatic gonorrhea
for White men.}

\item{ct.sympt.prob.tx.B}{Probability of treatment for symptomatic chlamydia
for Black men.}

\item{ct.sympt.prob.tx.W}{Probability of treatment for symptomatic chlamydia
for White men.}

\item{gc.asympt.prob.tx.B}{Probability of treatment for asymptomatic gonorrhea
for Black men.}

\item{gc.asympt.prob.tx.W}{Probability of treatment for asymptomatic gonorrhea
for White men.}

\item{ct.asympt.prob.tx.B}{Probability of treatment for asymptomatic chlamydia
for Black men.}

\item{ct.asympt.prob.tx.W}{Probability of treatment for asymptomatic chlamydia
for White men.}

\item{prep.sti.screen.int}{Interval in days between STI screening at PrEP visits.}

\item{prep.sti.prob.tx}{Probability of treatment given positive screening during
PrEP visit.}

\item{prep.continue.stand.tx}{Logical, if \code{TRUE} will continue standard
STI treatment of symptomatic cases even after PrEP initiation.}

\item{sti.cond.eff}{Relative risk of STI infection from anal sex when a condom is
used properly (biological efficacy).}

\item{sti.cond.fail.B}{Condom failure rates for STI for Black MSM, as a reduction
in the cond.eff parameter.}

\item{sti.cond.fail.W}{Condom failure rates for STI for White MSM, as a reduction
in the cond.eff parameter.}

\item{hiv.rgc.rr}{Relative risk of HIV infection given current rectal gonorrhea.}

\item{hiv.ugc.rr}{Relative risk of HIV infection given current urethral gonorrhea.}

\item{hiv.rct.rr}{Relative risk of HIV infection given current rectal chlamydia.}

\item{hiv.uct.rr}{Relative risk of HIV infection given current urethral chlamydia.}

\item{hiv.dual.rr}{Additive proportional risk, from 0 to 1, for HIV infection
given dual infection with both gonorrhea and chlamydia.}

\item{...}{Additional arguments passed to the function.}

\item{cond.fail.B}{Condom failure rates for HIV for Black/White MSM, as a reduction
in the cond.eff parameter (vector of length 2).}

\item{prep.adhr.dist.B}{Proportion of black men who are low, medium, and high
adherent to PrEP.}

\item{prep.adhr.dist.W}{Proportion of white men who are low, medium, and high
adherent to PrEP.}
}
\value{
A list object of class \code{param_msm}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{msm}
